% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_output.R
\name{coef.hglm}
\alias{coef.hglm}
\title{Extraction of hglm model coefficients}
\usage{
\method{coef}{hglm}(object, ...)
}
\arguments{
\item{object, }{fitted model output of a hiper_glm() function call (class hglm)}

\item{..., }{potential additional arguments required for generic coef}
}
\value{
Named vector containing coefficient point estimates
}
\description{
Extraction of hglm model coefficients
}
\details{
Makes point estimates of hglm coefficients available.
}
