% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiper_glm.R
\name{hiper_glm}
\alias{hiper_glm}
\title{GLM for high-dimensional scenarios}
\usage{
hiper_glm(design, outcome)
}
\arguments{
\item{design, }{design matrix for the model}

\item{outcome, }{outcome vector/matrix for the model}
}
\value{
S3 object containing output of fitting model
}
\description{
GLM for high-dimensional scenarios
}
\details{
Fits a GLM to given modelling data of high dimension and returns output object
}
