% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_output.R
\name{print.hglm}
\alias{print.hglm}
\title{Create and display curated output from hglm}
\usage{
\method{print}{hglm}(x, ...)
}
\arguments{
\item{x, }{fitted model output of a hiper_glm() function call (class hglm)}

\item{..., }{potential additional arguments required for generic print}
}
\value{
String displayed
}
\description{
Create and display curated output from hglm
}
\details{
Generates informative output from fitted hglm
}
