% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_output.R
\name{vcov.hglm}
\alias{vcov.hglm}
\title{Extraction of variance-covariance of hglm model coefficients}
\usage{
\method{vcov}{hglm}(object, ...)
}
\arguments{
\item{object, }{fitted model output of a hiper_glm() function call (class hglm)}

\item{..., }{potential additional arguments required for generic vcov}
}
\value{
Labelled variance-covariance matrix
}
\description{
Extraction of variance-covariance of hglm model coefficients
}
\details{
Makes variance-covariance matrix of hglm coefficient estimates available
}
